requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
pillow==10.3.0
pyodbc==5.1.0
screeninfo==0.8.1
